fix: cast vec_ids to int64 for compatibility in search function

+++ IVF.py
@@ -191,7 +191,7 @@ def search(vec_db, query_vector, k=5, batch_size=500):
 
         vecs = vecs / (np.linalg.norm(vecs, axis=1, keepdims=True) + 1e-12)
 
-        for vid, vec in zip(vec_ids, vecs):
+        for vid, vec in zip(vec_ids.astype(np.int64), vecs):
             score = cal_score(query_vector, vec)
 
             item = (score, -vid)
